完成用户的权限控制

+++ src/main/java/com/lele/apps/cms/service/impl/CategoryServiceImpl.java
@@ -2,9 +2,12 @@ package com.lele.apps.cms.service.impl;
 
 import com.lele.apps.cms.bean.Category;
 import com.lele.apps.cms.bean.CategoryExample;
+import com.lele.apps.cms.bean.extend.CategoryExtend;
 import com.lele.apps.cms.dao.CategoryMapper;
+import com.lele.apps.cms.dao.extend.CategoryExtendMapper;
 import com.lele.apps.cms.service.ICategoryService;
 import com.lele.apps.cms.utils.CustomerException;
+import com.lele.apps.cms.utils.MessageUtil;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -25,6 +28,9 @@ public class CategoryServiceImpl implements ICategoryService {
     @Resource
     private CategoryMapper categoryMapper;
     
+    @Resource
+    private CategoryExtendMapper categoryExtendMapper;
+    
     
     @Override
     public List<Category> findAll () {
@@ -33,27 +39,27 @@ public class CategoryServiceImpl implements ICategoryService {
     
     @Override
     public void saveOrUpdate (Category category) throws CustomerException {
+        //判空处理
+        if (category.getName() == null || category.getName().equals("")) {
+            throw new CustomerException("栏目名称不能为空");
+        }
         //判断是插入还是更新
         if (category.getId() != null) {
             //保存操作
             categoryMapper.updateByPrimaryKey(category);
-        }else {
-            
-            //判空处理
-            if(category==null||category.equals("")||category.getName()==null){
-                throw new CustomerException("栏目名称不能为空");
-            }
-            
-            //判断栏目名称是不是和已有的重复
-            CategoryExample categoryExample = new CategoryExample();
-            categoryExample.createCriteria().andNameEqualTo(category.getName());
-            List<Category> list = categoryMapper.selectByExample(categoryExample);
-            if(list.size()>0)
-                throw  new CustomerException("栏目已经存在");
-    
-        //    插入操作
-            categoryMapper.insert(category);
+            return;
         }
+        
+        
+        //判断栏目名称是不是和已有的重复
+        CategoryExample categoryExample = new CategoryExample();
+        categoryExample.createCriteria().andNameEqualTo(category.getName());
+        List<Category> list = categoryMapper.selectByExample(categoryExample);
+        if (list.size() > 0)
+            throw new CustomerException("栏目已经存在");
+        
+        //    插入操作
+        categoryMapper.insert(category);
         
         
     }
@@ -62,7 +68,7 @@ public class CategoryServiceImpl implements ICategoryService {
     public void deleteById (Long id) throws CustomerException {
         
         Category category = categoryMapper.selectByPrimaryKey(id);
-        if(category==null)
+        if (category == null)
             throw new CustomerException("该栏目不存在");
         
         categoryMapper.deleteByPrimaryKey(id);
@@ -70,9 +76,23 @@ public class CategoryServiceImpl implements ICategoryService {
     
     @Override
     public void batchDelete (Long[] ids) throws CustomerException {
-        for(Long id:ids){
+        if (ids.length == 0) {
+            throw new CustomerException("请选择要删除的栏目");
+        }
+        
+        for (Long id : ids) {
             categoryMapper.deleteByPrimaryKey(id);
         }
+    }
+    
+    /**
+     * 查询所有的栏目。包括父栏目
+     *
+     * @return
+     */
+    @Override
+    public List<CategoryExtend> findAllIncludeParent () {
+        return categoryExtendMapper.selectAllIncludeParent();
     }
 }
